# src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mobile Money Bridge API",
    "version": "1.0.0",
    "description": "API for bridging mobile money providers (MTN MoMo, Airtel Money, Orange Money) with the Stellar blockchain network.\n\nEnables low-cost cross-border payments and remittances across Africa by connecting mobile money wallets to Stellar's decentralized network.\n\n## Authentication\n\nMost endpoints require a JWT bearer token obtained via `POST /api/auth/login`:\n```\nAuthorization: Bearer <token>\n```\n\nAdmin endpoints require an API key:\n```\nX-API-Key: <key>\n```\n\n## Key Concepts\n\n- **Deposit**: Mobile money fiat (XAF, XOF, etc.) → Stellar tokens (USDC, XLM)\n- **Withdraw**: Stellar tokens → Mobile money fiat in the destination country\n- **Settlement**: Stellar network handles cross-border settlement in ~5 seconds at minimal cost\n- **KYC**: Tiered identity verification (Unverified, Basic, Full) determines daily transaction limits",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/sublime247/mobile-money/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.mobile-money-bridge.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token obtained via `POST /api/auth/login`.\n\nInclude as: `Authorization: Bearer <token>`\n\nTokens expire after 24 hours. Use `POST /api/auth/refresh` with a valid refresh token to obtain a new access token without re-authenticating."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for admin-level operations.\n\nInclude as: `X-API-Key: <key>`\n\nAdmin API keys are generated via the admin dashboard. They grant full access to admin endpoints including user management, system configuration, and transaction oversight."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned on failed API requests.",
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code or short description.",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message with details.",
            "example": "Amount must be a positive number"
          }
        }
      },
...
      "TransactionListResponse": {
        "type": "object",
        "description": "Paginated list of transactions.",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionDetail" }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "List transactions with pagination and filters",
        "description": "Returns a paginated list of transactions for the authenticated user. Supports cursor-based and offset-based pagination. Filterable by status, type, provider, and date range.",
        "operationId": "listTransactions",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of transactions to return (1-100).",
            "schema": { "type": "integer", "maximum": 100, "example": 20 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of transactions to skip (offset-based pagination).",
            "schema": { "type": "integer", "minimum": 0, "example": 0 }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by transaction status.",
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "completed", "failed", "cancelled"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by transaction type.",
            "schema": {
              "type": "string",
              "enum": ["deposit", "withdraw"]
            }
          },
          {
            "name": "provider",
            "in": "query",
            "description": "Filter by mobile money provider.",
            "schema": {
              "type": "string",
              "enum": ["mtn", "airtel", "orange"]
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Cursor for backward pagination. Returns transactions created before this cursor ID.",
            "schema": { "type": "string" }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Cursor for forward pagination. Returns transactions created after this cursor ID.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of transactions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vaults": {
      "post": {
        "tags": ["Vaults"],
        "summary": "Create a new savings vault",
        "description": "Creates a new goal-based savings vault. Vaults hold Stellar assets and allow users to set savings goals, track progress, and earn yield on deposited funds.",
        "operationId": "createVault",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVaultRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vault created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "name": { "type": "string" },
                        "targetCurrency": { "type": "string" },
                        "balance": { "type": "string", "example": "0.00" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "409": {
            "description": "A vault with this name already exists.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "get": {
        "tags": ["Vaults"],
        "summary": "List all vaults for the authenticated user",
        "description": "Returns all savings vaults belonging to the authenticated user, with current balances and goal progress.",
        "operationId": "listVaults",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "includeInactive",
            "in": "query",
            "description": "If 'true', includes vaults that have been deactivated or fully withdrawn.",
            "schema": { "type": "string", "enum": ["true", "false"], "example": "false" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vaults.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
    "/graphql": {
      "post": {
        "tags": ["GraphQL"],
        "summary": "GraphQL API endpoint",
        "description": "GraphQL endpoint supporting queries, mutations, and real-time subscriptions.\n\nSee the GraphQL playground at /graphql in development mode for schema exploration and interactive query building.",
        "operationId": "graphql",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "description": "GraphQL query string." },
                  "variables": { "type": "object", "description": "Optional query variables." },
                  "operationName": { "type": "string", "description": "Optional operation name." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GraphQL response (may contain errors or data).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "object" },
                    "errors": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/queues": {
      "get": {
        "tags": ["Admin"],
        "summary": "BullMQ queue monitoring dashboard",
        "description": "Access the BullMQ admin dashboard for monitoring job queues, retrying failed jobs, and inspecting queue metrics. Requires admin API key.",
        "operationId": "adminQueues",
        "security": [{ "apiKey": [] }],
        "responses": {
          "200": {
            "description": "Queue dashboard HTML."
          },
          "401": {
            "description": "Unauthorized — valid API key required.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/reconciliation": {
      "get": {
        "tags": ["Admin"],
        "summary": "Provider reconciliation report",
        "description": "Returns reconciliation data comparing internal transaction records with mobile money provider statements. Used for financial auditing and discrepancy detection.",
        "operationId": "getReconciliation",
        "security": [{ "apiKey": [] }],
        "responses": {
          "200": {
            "description": "Reconciliation report.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/api/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Transaction statistics",
        "description": "Returns aggregate transaction statistics including volume, counts by status, provider breakdowns, and time-series data.",
        "operationId": "getStats",
        "security": [{ "apiKey": [] }],
        "responses": {
          "200": {
            "description": "Statistics data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}